+++ .github/workflows/validate_redirects.yml
@@ -1,4 +1,4 @@
-name: ValidateRedirects
+name: Validate Redirects
 on:
   pull_request:
     branches: [main]
@@ -8,7 +8,7 @@ env:
 permissions:
   contents: read
 jobs:
-  CheckConsoleErrors:
+  ValidateRedirects:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout repository
@@ -19,22 +19,14 @@ jobs:
           node-version: 20.x
       - name: Install Dependencies
         run: yarn
-      - name: Run Build
-        run: yarn build:release
-        env:
-          NODE_OPTIONS: --max_old_space_size=4096
-      - name: Run Server
-        run: |
-          python -m http.server 3000 -d ${{ env.BUILD_DIR }} &
-          sleep 5
       - name: Validate redirects
         id: redirects
         uses: actions/github-script@60a0d83039c74a4aee543508d2ffcb1c3799cdea # v7.0.1
         with:
           result-encoding: string
           script: |
-            const { redirects } = require('./tasks/validate_redirects');
+            const { invalidRedirects } = require('./.github/workflows/scripts/validate-redirects.js');
             return await invalidRedirects();
-      - name: Fail if console errors have been found
+      - name: Fail if any invalid redirects have been found
         if: ${{ steps.redirects.outputs.result }}
-        run: exit 1
+        run: exit 1 && echo ${{ steps.redirects.outputs.result }}
